API_tests/requirements.txt
requests
pytest
Faker
allure-pytest